2.1.0: ffmpeg: monitoring, file/frame is now 32 bit instead of 64.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.1.0@2359 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

Former-commit-id: 1cc5c536e472e755fa285370f94bad98583aa0bb

+++ modules/ffmpeg/producer/ffmpeg_producer.cpp
@@ -151,8 +151,8 @@ public:
 		*event_subject_ << monitor::event("profiler/time")		% frame_timer.elapsed() % (1.0/format_desc_.fps)					
 						<< monitor::event("file/time")			% monitor::duration(file_frame_number()/fps_) 
 																% monitor::duration(file_nb_frames()/fps_)
-						<< monitor::event("file/frame")			% static_cast<int64_t>(file_frame_number())
-																% static_cast<int64_t>(file_nb_frames())
+						<< monitor::event("file/frame")			% static_cast<int32_t>(file_frame_number())
+																% static_cast<int32_t>(file_nb_frames())
 						<< monitor::event("file/fps")			% fps_
 						<< monitor::event("file/video/mode")	% u8(print_mode())
 						<< monitor::event("file/video/codec")	% (video_decoder_ ? u8(video_decoder_->print()) : "n/a")
